refactor(appkit): retire toPluginWithInstance; consolidate on fromPlugin
With `fromPlugin` landed, there is no longer a reason for core plugin
factories to eagerly construct an instance at module scope just to
expose `.toolkit()` on the factory return value. `fromPlugin(factory)`
reads `factory.pluginName` synchronously and defers resolution to
`AgentsPlugin.setup()`, which calls `.toolkit()` as an instance method
on the real plugin.

Changes:

- Swap `toPluginWithInstance(X, ["toolkit"] as const)` to `toPlugin(X)`
  in analytics, files, genie, and lakebase. `.toolkit()` remains as an
  instance method on each Plugin class — it's still called by the
  markdown loader, by `AgentsPlugin.applyAutoInherit`, and by the new
  `resolveToolkitFromProvider` helper.
- Delete `toPluginWithInstance` from
  `packages/appkit/src/plugin/to-plugin.ts` and its re-export from the
  `plugin/` barrel. The `pluginName` stamp on `toPlugin` stays — that's
  what `fromPlugin` reads.
- Revert the PR 5e plumbing it enabled: drop `instance?: BasePlugin`
  from `OptionalConfigPluginDef` in `packages/shared/src/plugin.ts`,
  undo the `preparePlugins` forwarding and the `pluginData.instance`
  reuse branch in `packages/appkit/src/core/appkit.ts`, and delete the
  instance-reuse test. Nothing produces an `instance` field anymore.
- Extract the duplicated provider-resolution logic (previously
  copy-pasted between `AgentsPlugin` and `run-agent`) into a single
  shared helper at
  `packages/appkit/src/plugins/agents/toolkit-resolver.ts`. Both the
  `fromPlugin` marker resolution, the markdown auto-inherit path, and
  the standalone `runAgent` executor now go through it.
- Rewrite the `describe("toolkit()")` block in analytics.test.ts to
  construct `AnalyticsPlugin` directly instead of reaching for the
  factory return value; the behavior being tested lives on the class.
- Docs: drop the "Using `.toolkit()` directly (advanced)" subsection
  from `agents.md` and the "before / after" duplication from the
  migration guide. `fromPlugin` is now the only documented path.

No public-API regression for the new shape: every in-tree migration
already went through `fromPlugin` in PR 5f. Users who were still
spreading `.toolkit()` at module scope (not possible in-tree, but the
escape hatch was documented) need to switch to `fromPlugin(factory,
opts)` — same options, one less object-lifetime concern.

Net: ~370 lines removed, one canonical factory abstraction, one shared
resolver.
