+++ packages/appkit/src/core/appkit.ts
@@ -76,9 +76,7 @@ export class AppKit<TPlugins extends InputPluginMap> {
       name,
       ...extraData,
     };
-    // If the factory eagerly constructed an instance (via
-    // `toPluginWithInstance`), reuse it; otherwise construct now.
-    const pluginInstance = pluginData.instance ?? new Plugin(baseConfig);
+    const pluginInstance = new Plugin(baseConfig);
 
     if (typeof pluginInstance.attachContext === "function") {
       pluginInstance.attachContext({
@@ -231,11 +229,9 @@ export class AppKit<TPlugins extends InputPluginMap> {
   ) {
     const result: InputPluginMap = {};
     for (const currentPlugin of plugins) {
-      const instance = (currentPlugin as { instance?: BasePlugin }).instance;
       result[currentPlugin.name] = {
         plugin: currentPlugin.plugin,
         config: currentPlugin.config as Record<string, unknown>,
-        instance,
       };
     }
     return result;
