+++ packages/appkit/src/plugins/analytics/tests/analytics.test.ts
@@ -610,10 +610,9 @@ describe("Analytics Plugin", () => {
   });
 
   describe("toolkit()", () => {
-    test("factory exposes a toolkit() method producing ToolkitEntry records", () => {
-      const pluginData = analytics({});
-      expect(typeof pluginData.toolkit).toBe("function");
-      const entries = pluginData.toolkit();
+    test("produces ToolkitEntry records keyed by the plugin name", () => {
+      const plugin = new AnalyticsPlugin({ name: "analytics" });
+      const entries = plugin.toolkit();
       expect(Object.keys(entries)).toContain("analytics.query");
       const entry = entries["analytics.query"];
       expect(entry.__toolkitRef).toBe(true);
@@ -621,9 +620,9 @@ describe("Analytics Plugin", () => {
       expect(entry.localName).toBe("query");
     });
 
-    test("toolkit() respects prefix and only options", () => {
-      const pluginData = analytics({});
-      const entries = pluginData.toolkit({ prefix: "", only: ["query"] });
+    test("respects prefix and only options", () => {
+      const plugin = new AnalyticsPlugin({ name: "analytics" });
+      const entries = plugin.toolkit({ prefix: "", only: ["query"] });
       expect(Object.keys(entries)).toEqual(["query"]);
     });
   });
